# Improve performance of first/last_possible_days check