Apparently there was a better way to do this. 

Corrects the performance problem we found : 
- for a given instance we used ~3 to 4 seconds to solve before implementing first/last_possible_days check
- we needed ~10 seconds after implementing it (which correspond to 16 seconds on beta server)

This edit allows to solve this VRP in ~ 5 seconds. We do increase computation time little but we are make lot of additional checks in order to respect new provided constraints so this is consistent for me.